fixed float casting bugs

+++ vowpalwabbit/allreduce.h
@@ -62,8 +62,8 @@ template <class T, void (*f)(T&, const T&)> void addbufs(T* buf1, const T* buf2,
 void all_reduce_init(const string master_location, const size_t unique_id, const size_t total, const size_t node, node_socks& socks);
 
 template <class T> void pass_up(char* buffer, size_t left_read_pos, size_t right_read_pos, size_t& parent_sent_pos, socket_t parent_sock, size_t n) {
-  size_t my_bufsize = min(ar_buf_size, ((int)(floor(left_read_pos/((float)sizeof(T)))*sizeof(T)) - parent_sent_pos));
-  my_bufsize = min(my_bufsize, ((int)(floor(right_read_pos/((float)sizeof(T)))*sizeof(T)) - parent_sent_pos));
+  size_t my_bufsize = min(ar_buf_size, ((int)(floor(left_read_pos/((double)sizeof(T)))*sizeof(T)) - parent_sent_pos));
+  my_bufsize = min(my_bufsize, ((int)(floor(right_read_pos/((double)sizeof(T)))*sizeof(T)) - parent_sent_pos));
 
   if(my_bufsize > 0) {
     //going to pass up this chunk of data to the parent
